+++ projects/hip-tests/catch/unit/memory/hipHostRegister.cc
@@ -931,11 +931,11 @@ TEMPLATE_TEST_CASE("Unit_hipHostRegister_Flags", "", int, float, double) {
   have a definition in the headers */
   /* hipHostRegisterIoMemory is a valid flag but requires access to I/O mapped
   memory to be tested */
-  FlagType flags = GENERATE(FlagType{hipHostRegisterDefault, true},
-    FlagType{hipHostRegisterPortable, true},
-    FlagType{0x08, true},
-    FlagType{hipHostRegisterPortable | hipHostRegisterMapped, true},
-    FlagType{hipHostRegisterPortable | hipHostRegisterMapped | 0x08, true},
+  FlagType flags = GENERATE(
+      FlagType{hipHostRegisterDefault, true}, FlagType{hipHostRegisterPortable, true},
+      FlagType{hipHostRegisterIoMemory, true},
+      FlagType{0x08, true}, FlagType{hipHostRegisterPortable | hipHostRegisterMapped, true},
+      FlagType{hipHostRegisterPortable | hipHostRegisterMapped | 0x08, true},
 #if (HT_AMD == 1) && (HT_LINUX == 1)
     FlagType{hipExtHostRegisterUncached, true},
     FlagType{hipHostRegisterPortable | hipHostRegisterMapped | hipExtHostRegisterUncached, true},