+++ projects/clr/CHANGELOG.md
@@ -10,6 +10,7 @@ Full documentation for HIP is available at [rocm.docs.amd.com](https://rocm.docs
 * Support for the `hipMemAllocationTypeUncached` flag, enabling developers to allocate uncached memory. This flag is now supported in the following APIs:
     - `hipMemGetAllocationGranularity` determines the recommended allocation granularity for uncached memory.
     - `hipMemCreate` allocates memory with uncached properties.
+    - `hipHostRegister` now supports hipHostRegisterIoMemory flag
 
 ### Resolved issues